pyproject.toml
[project]
name = "agenticbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-aiplatform>=1.121.0",
    "google-cloud-bigquery>=3.38.0",
    "google-generativeai>=0.8.5",
    "pydantic>=2.12.3",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
